+++ app/checkout/payment/page.tsx
@@ -1,19 +1,19 @@
 'use client';
 
+import { useRouter } from 'next/navigation';
 import { Container } from "@/components";
 import PaymentForm from "@/components/forms/PaymentForm";
-import { useCart } from "@/context/CartContext";
-import { ITEM_PRICE } from "@/lib/config";
 
 export default function PaymentPage() {
-    const { cartItems, getCartTotal, updateQuantity, removeFromCart } = useCart();
-    const formattedCartTotal = `Total: £${getCartTotal() * ITEM_PRICE}.00`
+    const router = useRouter();
+    const handleSuccess = (orderId: string) => router.push(`/checkout/success?orderId=${orderId}`);
 
     return (
         <Container containerSize="lg">
             <h2 className="text-xl font-semibold mb-8">Checkout</h2>
-            <PaymentForm />
+            <PaymentForm onSuccess={handleSuccess} />
         </Container>
+
     );
 
 }